feat : JuiceMaker에서 Juice타입 분리
JuiceMaker에선 extension이용하여 확장
