+++ JuiceMaker/JuiceMaker/Model/JuiceMaker.swift
@@ -10,42 +10,30 @@ enum JuiceMakeError: Error {
     case lackOfStock
 }
 
-struct JuiceMaker {
-    private var fruitStore = FruitStore.fruitStore
-    
-    enum Juice: String {
-        case strawberryJuice = "딸기쥬스"
-        case bananaJuice = "바나나쥬스"
-        case pineappleJuice = "파인애플쥬스"
-        case kiwiJuice = "키위쥬스"
-        case mangoJuice = "망고쥬스"
-        case strawberryAndBananaJuice = "딸바쥬스"
-        case mangoAndKiwiJuice = "망키쥬스"
-        
-        var koreanName: String {
-            return self.rawValue
-        }
-        
-        var recipe: [(Fruit, Int)] {
-            switch self {
-            case .strawberryJuice:
-                return [(.strawberry, 16)]
-            case .bananaJuice:
-                return [(.banana, 2)]
-            case .pineappleJuice:
-                return [(.pineapple, 2)]
-            case .kiwiJuice:
-                return [(.kiwi, 3)]
-            case .mangoJuice:
-                return [(.mango, 3)]
-            case .strawberryAndBananaJuice:
-                return [(.strawberry, 10), (.banana, 1)]
-            case .mangoAndKiwiJuice:
-                return [(.mango, 2), (.kiwi, 1)]
-                
-            }
+extension Juice {
+    var recipe: [(Fruit, Int)] {
+        switch self {
+        case .strawberryJuice:
+            return [(.strawberry, 16)]
+        case .bananaJuice:
+            return [(.banana, 2)]
+        case .pineappleJuice:
+            return [(.pineapple, 2)]
+        case .kiwiJuice:
+            return [(.kiwi, 3)]
+        case .mangoJuice:
+            return [(.mango, 3)]
+        case .strawberryAndBananaJuice:
+            return [(.strawberry, 10), (.banana, 1)]
+        case .mangoAndKiwiJuice:
+            return [(.mango, 2), (.kiwi, 1)]
+            
         }
     }
+}
+
+struct JuiceMaker {
+    private var fruitStore = FruitStore.fruitStore
     
     func orderAndCheck(_ juice: Juice) -> Bool {
         do {
